Fix bug in dto for storing current temperature
Using forEachIndexed did not allow the application to break out
of the loop once the condition was met. Refactoring the code to be a
generic for loop enabled the use of the break keyword, meaning iteration
stopped.

Also, removed the need to convert weathercode to string and then back into integer
as it already was an integer.

Annotations added to other classes as some features require higher API levels.