Implement placeholder search location screen
TODO: The search screen currently uses a placeholder data class
inside of the viewmodel. To adapt this to your application,
you need to have a list of locations that you filter, and return
as a list. Afterwards, replace all list instances inside the
viewmodel with the new list. Also, you need to use this information
to change the API query so that you update the main weather display
with the new location.